fix: correct 64-bit nodeId truncation across MethodChannel
Matter nodeIds are unsigned 64-bit values. Two bugs caused them to
be silently corrupted on every round-trip through the Flutter
MethodChannel, making all post-commissioning commands (identify,
toggle, subscribe, etc.) fail:

1. MatterBridge.kt serialised nodeIds with .toInt(), discarding the
   upper 32 bits in commission results and all subscription/OTA events.

2. MainActivity.kt read nodeIds back with call.argument<Int>(), which
   returns null for any value that doesn't fit in a signed 32-bit int,
   causing the fallback 0L to be used instead of the real nodeId.

The generated commissioning nodeId (currentTimeMillis & 0xFFFFFFFF)
has bit 31 set roughly half the time, so ~50% of commissionings
produced a broken device that could not be communicated with.

Fix:
- MatterBridge.kt: pass nodeId as Long directly in all result maps
  and event payloads (commission result x2, subscription events x4,
  OTA progress events).
- MainActivity.kt: add a MethodCall.nodeIdArg() extension that handles
  both Int32 and Int64 wire values, replacing all 27 argument<Int>
  nodeId reads.

+++ android/app/src/main/kotlin/com/fluxhome/app/MatterBridge.kt
@@ -87,7 +87,7 @@ class MatterBridge(private val context: Context) {
             onUpdate = { nid, attrs ->
                 if (nid !in cancelledNodeIds) {
                     val payload = mutableMapOf<String, Any?>(
-                        "nodeId" to nid.toInt(),
+                        "nodeId" to nid,
                         "type"   to "update",
                     )
                     payload.putAll(attrs)
@@ -96,11 +96,11 @@ class MatterBridge(private val context: Context) {
             },
             onEstablished = { nid ->
                 if (nid !in cancelledNodeIds)
-                    emitDeviceState(mapOf("nodeId" to nid.toInt(), "type" to "established"))
+                    emitDeviceState(mapOf("nodeId" to nid, "type" to "established"))
             },
             onResubscribing = { nid, nextMs ->
                 if (nid !in cancelledNodeIds) {
-                    emitDeviceState(mapOf("nodeId" to nid.toInt(), "type" to "resubscribing",
+                    emitDeviceState(mapOf("nodeId" to nid, "type" to "resubscribing",
                                          "nextMs" to nextMs))
                     // SDK exponential backoff can grow to minutes — if it exceeds 30 s,
                     // the UDP socket is likely permanently broken.  Restart cleanly.
@@ -118,7 +118,7 @@ class MatterBridge(private val context: Context) {
             },
             onError = { nid, err ->
                 if (nid !in cancelledNodeIds) {
-                    emitDeviceState(mapOf("nodeId" to nid.toInt(), "type" to "error",
+                    emitDeviceState(mapOf("nodeId" to nid, "type" to "error",
                                          "message" to (err.message ?: "unknown")))
                     // For ICD / sleepy devices the initial connection may time out
                     // because the device is asleep.  Retry after 60 s — the device
@@ -171,7 +171,7 @@ class MatterBridge(private val context: Context) {
         )
         val deviceTypeId = readPrimaryDeviceType(commissionedNodeId)
         main.post {
-            result.success(mapOf("nodeId" to commissionedNodeId.toInt(), "deviceTypeId" to deviceTypeId))
+            result.success(mapOf("nodeId" to commissionedNodeId, "deviceTypeId" to deviceTypeId))
         }
     }
 
@@ -196,7 +196,7 @@ class MatterBridge(private val context: Context) {
         )
         val deviceTypeId = readPrimaryDeviceType(commissionedNodeId)
         main.post {
-            result.success(mapOf("nodeId" to commissionedNodeId.toInt(), "deviceTypeId" to deviceTypeId))
+            result.success(mapOf("nodeId" to commissionedNodeId, "deviceTypeId" to deviceTypeId))
         }
     }
 
@@ -516,7 +516,7 @@ class MatterBridge(private val context: Context) {
     private fun emitOtaProgress(nodeId: Long, phase: String, progress: Int?, message: String?) {
         val map = mutableMapOf<String, Any?>(
             "type"   to "otaProgress",
-            "nodeId" to nodeId.toInt(),
+            "nodeId" to nodeId,
             "phase"  to phase,
         )
         if (progress != null) map["progress"] = progress
